Layout_TestCase/http/tests/xmlhttprequest/upload-request-error-event-order-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/cross-site-progress-events.cgi. Origin http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
Test to validate the order in which the events are fired in case of a request error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS actualOrder is "rsdone,upload.onerror,upload.onloadend,onerror,onloadend,"
PASS successfullyParsed is true

TEST COMPLETE
